+++ hierarchical_sales_forecasting/b2b_revenue_forecasting/__init__.py
@@ -10,9 +10,9 @@ from b2b_revenue_forecasting.quota_cascader import (
 from b2b_revenue_forecasting.commit_reconciler import CommitReconciler
 from b2b_revenue_forecasting.pipeline_adjuster import PipelineAdjuster
 from b2b_revenue_forecasting.metric_spec import MetricSpec
-from b2b_revenue_forecasting.batch import cascade_many
+from b2b_revenue_forecasting.batch import cascade_many, route_targets
 
-__version__ = "0.13.0"
+__version__ = "0.14.0"
 
 __all__ = [
     "SalesHierarchy",
@@ -24,5 +24,6 @@ __all__ = [
     "PipelineAdjuster",
     "MetricSpec",
     "cascade_many",
+    "route_targets",
     "__version__",
 ]